+++ plugins/mockito-factories/src/main/kotlin/com/episode6/mxo2/plugins/mockito/factories/MockitoAutoFactory.kt
@@ -1,4 +1,4 @@
-package com.episode6.mxo2.plugins.mockito
+package com.episode6.mxo2.plugins.mockito.factories
 
 import com.episode6.mxo2.MockspressoBuilder
 import com.episode6.mxo2.MockspressoProperties
